refactor: isolate deprecated resource API in _LegacyResourcesMixin
self.resources, self.dialog_renderer and self.find_resource are kept
for one release while skills migrate to the spec-tools-backed path.
Move all three — and the _skill_resources_compat cache they share —
into ovos_workshop.skills._legacy_resources, which OVOSSkill inherits.

Ends the deprecation cycle in one removal: drop the mixin from
OVOSSkill's bases and delete the file. Nothing on OVOSSkill itself
references the legacy SkillResources path any more.
